+++ src/main/java/org/prgrms/kdt/DemoApplication.java
@@ -1,8 +1,10 @@
 package org.prgrms.kdt;
 
 import org.prgrms.kdt.io.Input;
+import org.prgrms.kdt.io.InputConsole;
 import org.prgrms.kdt.io.Output;
-import org.prgrms.kdt.model.VoucherProgram;
+import org.prgrms.kdt.io.OutputConsole;
+import org.prgrms.kdt.model.voucher.VoucherProgram;
 import org.prgrms.kdt.service.VoucherService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -12,8 +14,9 @@ public class DemoApplication {
 
     public static void main(String[] args) {
         var applicationContext = SpringApplication.run(DemoApplication.class, args);
-        Input input = new Input();
-        Output output = new Output();
+
+        Input input = new InputConsole();
+        Output output = new OutputConsole();
         VoucherService voucherService = applicationContext.getBean(VoucherService.class);
 
         new VoucherProgram(input, output, voucherService).run();
