+++ src/main/java/org/prgrms/kdt/io/Output.java
@@ -1,34 +1,16 @@
 package org.prgrms.kdt.io;
 
-import org.prgrms.kdt.model.Function;
-import org.prgrms.kdt.model.VoucherType;
+public interface Output {
 
-import java.text.MessageFormat;
+    void printFunctions();
 
-public class Output {
-    public void printFunctions() {
-        System.out.println("=== Voucher Program ===");
-        for (Function function : Function.values()) {
-            System.out.println(MessageFormat.format("Type {0}{1}.", function.name(), function.getExplain()));
-        }
-    }
+    void printInputFunctionError();
 
-    public void printInputFunctionError() {
-        System.out.println("Type right command");
-    }
+    void printVoucherType();
 
-    public void printVoucherType() {
-        System.out.println("Type number of voucher you make");
-        for (VoucherType voucherType : VoucherType.values()) {
-            System.out.println(MessageFormat.format("{0} : {1}", voucherType.name(), voucherType.getTypeNumber()));
-        }
-    }
+    void printVoucherListEmptyError();
 
-    public void printVoucherListEmptyError() {
-        System.out.println("voucher list is empty");
-    }
+    void printExitMessage();
 
-    public void printExitMessage() {
-        System.out.println("exit program ! bye :)");
-    }
+    void illegalArgumentExceptionMessage(IllegalArgumentException e);
 }
