+++ src/main/java/org/prgrms/kdt/io/OutputConsole.java
@@ -1,0 +1,44 @@
+package org.prgrms.kdt.io;
+
+import org.prgrms.kdt.model.function.Function;
+import org.prgrms.kdt.model.voucher.VoucherType;
+
+import java.text.MessageFormat;
+
+public class OutputConsole implements Output {
+    @Override
+    public void printFunctions() {
+        System.out.println("=== Voucher Program ===");
+        for (Function function : Function.values()) {
+            System.out.println(MessageFormat.format("Type {0}{1}.", function.name(), function.getExplain()));
+        }
+    }
+
+    @Override
+    public void printInputFunctionError() {
+        System.out.println("WRONG : Type right command\n");
+    }
+
+    @Override
+    public void printVoucherType() {
+        System.out.println("=== Type number of voucher you make ===");
+        for (VoucherType voucherType : VoucherType.values()) {
+            System.out.println(MessageFormat.format("{0} : {1}", voucherType.name(), voucherType.getTypeNumber()));
+        }
+    }
+
+    @Override
+    public void printVoucherListEmptyError() {
+        System.out.println("voucher list is empty !!\n");
+    }
+
+    @Override
+    public void printExitMessage() {
+        System.out.println("exit program ! bye :)");
+    }
+
+    @Override
+    public void illegalArgumentExceptionMessage(IllegalArgumentException e) {
+        System.out.println(MessageFormat.format("{0}\n", e.getMessage()));
+    }
+}
