--- src/main/java/org/prgrms/kdt/model/Voucher.java
@@ -1,8 +1,0 @@
-package org.prgrms.kdt.model;
-
-import java.util.UUID;
-
-public interface Voucher {
-    UUID getVoucherId();
-    long getDiscountAmount();
-}
