+++ src/main/java/org/prgrms/kdt/model/function/FunctionMapping.java
@@ -1,0 +1,18 @@
+package org.prgrms.kdt.model.function;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class FunctionMapping {
+    private static Map<String, Function> functionMap = new HashMap<>();
+
+    static {
+        functionMap.put("create", Function.create);
+        functionMap.put("list", Function.list);
+        functionMap.put("exit", Function.exit);
+    }
+
+    public static Function valueOf(String function) {
+        return functionMap.get(function);
+    }
+}
