+++ src/main/java/org/prgrms/kdt/model/voucher/Voucher.java
@@ -1,0 +1,9 @@
+package org.prgrms.kdt.model.voucher;
+
+import java.util.UUID;
+
+public interface Voucher {
+    UUID getVoucherId();
+
+    long getDiscountAmount();
+}
