+++ src/main/java/org/prgrms/kdt/repository/VoucherRepository.java
@@ -1,11 +1,12 @@
 package org.prgrms.kdt.repository;
 
-import org.prgrms.kdt.model.Voucher;
+import org.prgrms.kdt.model.voucher.Voucher;
 
 
 import java.util.Map;
 
 public interface VoucherRepository<K, V> {
     Voucher insert(Voucher voucher);
-    Map<K,V> getVoucherList();
+
+    Map<K, V> getVoucherList();
 }
