+++ src/test/java/org/prgrms/kdt/model/FunctionTest.java
@@ -1,6 +1,7 @@
 package org.prgrms.kdt.model;
 
 import org.junit.jupiter.api.Test;
+import org.prgrms.kdt.model.function.Function;
 import org.prgrms.kdt.service.VoucherService;
 
 import static org.mockito.Mockito.*;
@@ -13,7 +14,7 @@ class FunctionTest {
     void doFunction() {
         VoucherService voucherService = mock(VoucherService.class);
 
-        Boolean exit = Function.exit.doFunction(voucherService);
+        Boolean exit = Function.exit.execute(voucherService);
 
         assertThat(exit, is(true));
     }
